requirements.txt
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
bcrypt==4.0.1
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.22.0
cmdstanpy==1.2.5
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.11
dnspython==2.8.0
email-validator==2.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.116.1
fonttools==4.59.2
greenlet==3.2.4
h11==0.16.0
holidays==0.80
httptools==0.6.4
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
murmurhash==1.0.13
nltk==3.9.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
passlib==1.7.4
patsy==1.0.1
pillow==11.3.0
preshed==3.0.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.2
scipy==1.16.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
stanio==0.5.1
starlette==0.47.3
statsmodels==0.14.5
thinc==8.3.6
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.3
